+++ src/app/admin/admin-routing.module.ts
@@ -2,24 +2,44 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { AdminComponent } from './admin.component';
-import { AdminDashboardComponent, ManageProductsComponent, ManageOrdersComponent } from './components';
+import { AdminDashboardComponent, ManageProductsComponent, ManageOrdersComponent, ProductFormComponent } from './components';
+import { AuthGuard } from '../core';
+import { ProductResolveGuard } from '../products';
 
 const routes: Routes = [
   {
-    path: 'admin',
+    path: '',
     component: AdminComponent,
+    canActivate: [AuthGuard],
     children: [
       {
-        path: 'products',
-        component: ManageProductsComponent
-      },
-      {
-        path: 'orders',
-        component: ManageOrdersComponent
-      },
-      {
         path: '',
-        component: AdminDashboardComponent
+        canActivateChild: [AuthGuard],
+        children: [
+          {
+            path: 'products',
+            component: ManageProductsComponent
+          },
+          {
+            path: 'products/add',
+            component: ProductFormComponent,
+          },
+          {
+            path: 'products/edit/:productID',
+            component: ProductFormComponent,
+            resolve: {
+              product: ProductResolveGuard
+            }
+          },
+          {
+            path: 'orders',
+            component: ManageOrdersComponent
+          },
+          {
+            path: '',
+            component: AdminDashboardComponent
+          }
+        ]
       }
     ]
   }
@@ -29,4 +49,8 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class AdminRoutingModule { }
+export class AdminRoutingModule {
+  static components = [
+    AdminDashboardComponent, ManageProductsComponent, ManageOrdersComponent, ProductFormComponent
+  ];
+ }