+++ src/app/admin/admin.module.ts
@@ -1,14 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 import { AdminRoutingModule } from './admin-routing.module';
 
 import { AdminComponent } from './admin.component';
 
 @NgModule({
-  declarations: [AdminComponent.components, AdminComponent],
+  declarations: [AdminRoutingModule.components, AdminComponent],
   imports: [
     CommonModule,
+    FormsModule,
     AdminRoutingModule
   ]
 })