+++ src/app/admin/components/index.ts
@@ -1,3 +1,4 @@
 export * from './admin-dashboard/admin-dashboard.component';
 export * from './manage-products/manage-products.component';
 export * from './manage-orders/manage-orders.component';
+export * from './product-form/product-form.component';