+++ src/app/admin/components/manage-orders/manage-orders.component.ts
@@ -1,5 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 
+import { OrderService } from '../../../core';
+import { CartItem } from '../../../cart/models';
+
 @Component({
   selector: 'app-manage-orders',
   templateUrl: './manage-orders.component.html',
@@ -7,9 +10,18 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ManageOrdersComponent implements OnInit {
 
-  constructor() { }
+  orders: CartItem[][];
+  constructor(
+    private orderService: OrderService
+  ) { }
 
   ngOnInit() {
+    this.orders = this.orderService.getOrderList();
+  }
+
+  onDelete(deletingOrder: CartItem[]) {
+    this.orderService.removeOrder(deletingOrder);
+    this.orders = this.orderService.getOrderList();
   }
 
 }