+++ src/app/app.component.ts
@@ -2,6 +2,8 @@ import { Component, ViewChild, AfterViewInit, ElementRef } from '@angular/core';
 
 import { ConstantsServiceService } from './core';
 
+import { SpinnerService } from './widgets';
+
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
@@ -10,7 +12,9 @@ import { ConstantsServiceService } from './core';
 export class AppComponent implements AfterViewInit {
   @ViewChild('appTitle', { static: true }) titleRef: ElementRef;
 
-  constructor(private constantsServiceService: ConstantsServiceService) {}
+  constructor(
+    public spinnerService: SpinnerService,
+    private constantsServiceService: ConstantsServiceService) {}
 
   ngAfterViewInit() {
     (this.titleRef.nativeElement as HTMLElement).innerText = this.constantsServiceService.title;