+++ src/app/core/index.ts
@@ -1,4 +1,7 @@
 export * from './services/local-storage.service';
 export * from './services/config-options.service';
 export * from './services/constants-service.service';
+export * from './services/auth.service';
 export * from './services/generator.service';
+export * from './services/order.service';
+export * from './guards/auth.guard';