+++ src/app/layout/components/index.ts
@@ -1,2 +1,3 @@
 export * from './about/about.component';
 export * from './order/order.component';
+export * from './login/login.component';