+++ src/app/layout/components/order/order.component.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
+import { OrderService } from '../../../core';
+import { CartService } from '../../../cart/services';
+import { CartItem } from '../../../cart/models';
+
 @Component({
   selector: 'app-order',
   templateUrl: './order.component.html',
@@ -7,9 +11,18 @@ import { Component, OnInit } from '@angular/core';
 })
 export class OrderComponent implements OnInit {
 
-  constructor() { }
+  cartItemList: CartItem[] = [];
 
-  ngOnInit() {
+  constructor(
+    private cartService: CartService,
+    private orderService: OrderService
+  ) { }
+
+  ngOnInit(): void {
+    this.cartItemList = this.cartService.cartItemList;
   }
 
+  onConfirm(): void {
+    this.orderService.addOrder(this.cartItemList);
+  }
 }