+++ src/app/products/index.js
@@ -1,6 +1,6 @@
 export * from './components/product/product.component';
 export * from './components/product-list/product-list.component';
-
 export * from './models/product.model';
-
+export * from './services/product.service';
+export * from './guards';
 export * from './products.module';